+++ clients/Python-PyPy/run.sh
@@ -1,0 +1,4 @@
+set -ex
+
+cd /output
+pypy3 ./main.py "$@"